% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc_int.R
\name{mc_int}
\alias{mc_int}
\title{Monte Carlo Integration}
\usage{
mc_int(x_range, fun, B, seed = 123)
}
\arguments{
\item{x_range}{A numeric vector of length 2 defining the integration interval [a, b].}

\item{fun}{A character string representing the function to integrate (e.g., "x^2").}

\item{B}{Number of Monte Carlo samples (positive integer).}

\item{seed}{Random seed (default: 123).}
}
\value{
A list containing the estimated integral value (`I`) and variance (`var`).
}
\description{
Estimate the integral of a function over a specified interval using Monte Carlo methods.
}
\details{
Monte Carlo Integration
}
