% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.MCI.R
\name{plot.MCI}
\alias{plot.MCI}
\title{Plot Monte Carlo Integration Result}
\usage{
\method{plot}{MCI}(x, ...)
}
\arguments{
\item{x}{An object of class `MCI` from `mc_int`.}

\item{...}{Additional graphical parameters.}
}
\description{
Visualize the function and integration area.
}
